chore(form-summary): move files to __interal__ folder
