+++ src/components/form/form.spec.js
@@ -16,8 +16,11 @@ import {
   StyledFormFooter,
   StyledForm,
 } from "./form.style";
-import FormSummary from "./form-summary.component";
-import { StyledFormSummary, StyledInternalSummary } from "./form-summary.style";
+import FormSummary from "./__internal__/form-summary.component";
+import {
+  StyledFormSummary,
+  StyledInternalSummary,
+} from "./__internal__/form-summary.style";
 import Icon from "../icon";
 import Button from "../button";
 import { FieldsetStyle } from "../../__experimental__/components/fieldset/fieldset.style";
